Backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
joblib==1.3.2
mysqlclient==2.1.1
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
threadpoolctl==3.2.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
